# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * REST API for the Smart Governance Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from './complaintStatus';

export interface Complaint {
  id: number;
  reference: string;
  title: string;
  category: string;
  description: string;
  location: string;
  status: ComplaintStatus;
  citizenId: number;
  citizenName: string;
  departmentId: number;
  departmentName: string;
  /** @nullable */
  assignedStaffId: number | null;
  /** @nullable */
  assignedStaffName: string | null;
  /** @nullable */
  remarks: string | null;
  /** @nullable */
  resolution: string | null;
  createdAt: string;
  updatedAt: string;
}